requirements.txt
sasl
thrift
thrift-sasl
pyhive
python-dotenv
tabulate